timer-backend/pyproject.toml
[project]
name = "timer-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "httpx[http2]>=0.28.1",
    "motor>=3.7.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
